feat: refactor Sequencer to ShapeProcessor
